+++ .clang-format
@@ -4,3 +4,4 @@ Language: Cpp
 IndentWidth: 2
 ColumnLimit: 120
 AllowShortIfStatementsOnASingleLine: false
+SortIncludes: false
