+++ include/reactor-uc/logging.h
@@ -1,5 +1,6 @@
 #ifndef REACTOR_UC_LOGGING_H
 #define REACTOR_UC_LOGGING_H
+#include <sys/types.h>
 #include <inttypes.h>
 
 // The different verbosity levels supported
